Create perception.cpp
This node gets an image from "robot2/camera/rgb/image_raw" and shows it. It also publishes "5" periodically in "/area". It will be upgraded to convert the received image to HSV, threshold it, extract contours and get the biggest one, from which will extract the centroid, area and bounding box, publishing all of them.